fix(cloudzap): wrap cores correctly
By embedding the wrapped core (and not implementing Wrap), calling Wrap
on it would return the "root" core, effectively dropping all wrapped
cores.

This fixes an issue where no Error Reporting fields are appended to the
log entry when the logger is derived using `cloudzap.WithLoggerFields`
(which in turn calls `core.Wrap`).