feat: add method for accessing trace context from http
Currently the trace context is read from incoming metadata which is
only populated when the server is called via gRPC. This means there is
no way (see note 1) to read the trace context from a http server (as
incoming metadata is not populated.

---

This commit introduces a separate context.Context value which is populated
by middlewares and accessible for both HTTP and gRPC servers.

The old method is deprecated and replaced with Get/Set Context funcs.

---

Note 1: Technically you can access the trace context in a http server by
reading outgoing metadata, but that is not ideal.