fix(cloudmonitoring): demote stop Error log to Warn on InvalidArgument
When stopping the OpenTelemetry Cloud Monitoring exporter, it tries to
export metrics a final time. This may be too close in time to the
previous export, causing a failure of exporting too frequently:

> rpc error: code = InvalidArgument desc = One or more TimeSeries could
> not be written: One or more points were written more frequently than
> the maximum sampling period configured for the metric.

While it's not optimal to lose the last <=1 minute of metrics, it
shouldn't trigger an Error log as it is not really actionable. Thus,
demoting this log to Warn instead in this specific case.