fix: align keepalive settings between client and server
With the current client keep alive settings, the server will
send GOAWAY to the client, as it is pinging too often and doing so
without an active stream.

This is because the default values for the server enforcement policy is
currently set to max one ping per 5min and no pings without active
streams.

This commit
* increases the time between pings from a client to 1 minute, as this
  should be sufficient to avoid gateway reaping the connection.
* decreases the server min time enforcement policy to 30s (was
  defaulting to 5min)
* configure server to allow client to send pings even if no stream is
  active.