fix(cloudmonitoring): remove latency metrics
The exported histograms do not make sense. It seems like they are of
CUMULATIVE type instead of DELTA. This has to be fixed within the
opentelemetry-operations-go library.

As there might be gRPC metrics introduced in the
opentelemetry-go-contrib project (see [1]), let's not try to fix them
here right now but instead remove it and use future ones in that
library.

[1]: https://github.com/open-telemetry/opentelemetry-go-contrib/issues/2840