feat: enable collection of OpenCensus metrics
For some Google Cloud libraries, metrics are emitted that are useful to
understanding the operation of a service but they are using the
OpenCensus SDKs rather than the OpenTelemetry ones.

A notable example is GoogleCloudPlatform/cloud-sql-go-connector which is
used as a proxy when connecting Cloud Run instances to Cloud SQL. There
is a [suite of metrics](https://github.com/GoogleCloudPlatform/cloud-sql-go-connector/blob/v1.4.5/internal/trace/metrics.go)
that are particularly useful in monitoring the health of the proxy.

This change allows users to enable capturing OpenCensus formatted metrics using
the OpenTelemetry bridge shim. By default this will be disabled.

There is a downside to enabling this producer, mainly that since OpenCensus
telemetry uses globals, installing a bridge will result in telemetry collection
from all libraries that use OpenCensus, including some you may not expect,
such as the telemetry exporter itself.

More info can be found in the following documentation:
https://pkg.go.dev/go.opentelemetry.io/otel/bridge/opencensus#hdr-Migration_Guide