feat(cloudslog): bridge log/slog with cloudzap
A new empty package is added: cloudslog that is intended to contain
handler(s) and funtions related to the new std lib slog package.

This change also replaces the default log/slog logger to a one that
instead logs to our default zap logger. This is useful if you use libs
or sdks that do not rely on cloudrunner and/or zap.