feat!: upgrade to latest stable otel API v1.21.0
otel uses a little bit of a strange versioning schema, where the
contracts are versioned under otel/semconv/{VERSION}, but the package
can be upgraded to several versions ahead of the contract. This has
worked fine for our use-cases, but we recently ran into a problem where
we go the following error on application startup - after a dependabot
upgrade came along:
```
cannot merge resource due to conflicting Schema URL
```
This led down a rabbit hole, eventually finding out this quirk in otel
usage.

BREAKING CHANGE: This upgrades the version of the otel contract being
used. This is possibly a breaking change so care must be taken so that
other OTEL packages are upgraded to v1.21 simultaneously so there aren't
further breakages.