feat(cloudotel): call reader.Shutdown instead of exporter.Shutdown
The periodic reader collects and exports metrics previously recorded at
a given interval.

In order to make sure we do not miss recorded metrics that have not yet
been exported at shutdown, we need to call the reader's Shutdown method
which performs a last collect and export.

It also performs a shutdown of the exporter so we don't need to do that
ourselves.

relates to [PE-910](https://einride.atlassian.net/browse/PE-910)