feat: on validate command, validate that all secrets are tagged
This loops through all environments defined in the cloud run yaml file.
If it contains secretKeyRef, (i.e referencing a GCP Secret Manager
Secret) it will go through all fields in the Golang Config to find the
matching environment name field.

If that field is not tagged as secret, we return an error.

This is only run on `-validate` but is a breaking change for that command.