feat(cloudrequsetlog): initial slog support
Step 1 is to change the public API in a backwards-compatible way so that
both slog and zap logger fields/attributes are supported. This is part
of step 1.

Step 2 is to change the underlying representation from zap to slog -
this will come in a follow up.

The conversion code is copied from the stdlib and the zap library.