fix(cloudrequestlog): use slog attrs instead of zap fields
The cloudrequestlog middleware uses a slog logger to log, but users
adding slog attrs, get converted to zap fields, and then back to slog
attrs when logged.

This changes so we store slog attrs, and convert zap fields to slog
attrs instead of the other way around.

This worked well in most cases when you used primitive types, but
if you used grouped values and such it didn't log that properly.

Note you can still provide zap fields to the cloudrequestlog for
compatibility reasons.