feat(cloudotel): allow partial and conflicting otel resources
When some OTEL dependencies are not synchronized, merging issues occur
and makes cloudrunner-go's cloudotel package fail on startup.

Partial resources and conflicting resource schema URLs are not always
(probably seldom?) a real problem that needs to stop a deploy. If this
causes telemetry issues, the warning can be acted upon to fix it.

I know this is quite opinionated, if there's not a consensus on this
then maybe I should work on making it opt-in somehow.